Make member functions work, partly implement boat options and functions

+++ BoatClub/src/model/Member.java
@@ -22,7 +22,7 @@ public class Member {
 	@XmlElement(name = "MemberName")
 	private String name;
 	@XmlElement(name = "PersonalNum")
-	private int personnumber;
+	private double personnumber;
 	@XmlElement(name = "MemberID")
 	private int memberId;
 	@XmlElement(name = "BoatNum")
@@ -30,7 +30,7 @@ public class Member {
 	@XmlElement(name = "BoatList")
 	private List<Boat> boatList = new ArrayList<Boat>();
 
-	public Member(String name, int personnumber, int memberId) throws ParseException {
+	public Member(String name, double personnumber, int memberId) throws ParseException {
 		this.name = name;
 		this.personnumber = personnumber;
 		this.memberId = memberId;
@@ -48,7 +48,7 @@ public class Member {
 		this.name = name;
 	}
 
-	public int getPersonnumber() {
+	public double getPersonnumber() {
 		return personnumber;
 	}
