+++ BoatClub/src/model/MemberCatalog.java
@@ -32,12 +32,12 @@ public class MemberCatalog {
 		this.memberList = memberCatalog;
 	}
 	
-	public void addMember(String name, int personnumber) throws ParseException {
+	public void addMember(String name, double personnumber) throws ParseException {
 		int id = getMemberList().size() + 1;
 		this.memberList.add(new Member(name, personnumber,  id));
 	}
 
-	public void updateMember(Member currentMember, String name, int personnumber) throws ParseException { // FIXME not sure if this is enough
+	public void updateMember(Member currentMember, String name, double personnumber) throws ParseException { // FIXME not sure if this is enough
 		memberList.remove(currentMember);
 		int id = getMemberList().size() + 1;
 		memberList.add(new Member(name, personnumber, id));
